+++ src/zepben/evolve/model/cim/iec61970/base/wires/load_break_switch.py
@@ -1,0 +1,17 @@
+#  Copyright 2023 Zeppelin Bend Pty Ltd
+#
+#  This Source Code Form is subject to the terms of the Mozilla Public
+#  License, v. 2.0. If a copy of the MPL was not distributed with this
+#  file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+from __future__ import annotations
+
+from zepben.evolve.model.cim.iec61970.base.wires.protected_switch import ProtectedSwitch
+
+__all__ = ["LoadBreakSwitch"]
+
+
+class LoadBreakSwitch(ProtectedSwitch):
+    """A mechanical switching device capable of making, carrying, and breaking currents under normal operating
+    conditions. """
+    pass