+++ test/busbranch/data/single_branch_common_lines_network.py
@@ -5,7 +5,8 @@
 #  file, You can obtain one at https://mozilla.org/MPL/2.0/.
 from busbranch.data.creators import _create_per_length_sequence_impedance
 from network_fixtures import create_terminal, create_terminals
-from zepben.evolve import NetworkService, AcLineSegment, Breaker
+from zepben.evolve import NetworkService, AcLineSegment
+from zepben.evolve.model.cim.iec61970.base.wires.breaker import Breaker
 
 
 def single_branch_common_lines_network(sw_is_open: bool) -> NetworkService: