+++ test/cim/iec61970/base/wires/test_jumper.py
@@ -7,7 +7,7 @@ from hypothesis import given
 
 from cim.iec61970.base.wires.test_switch import verify_switch_constructor_default, verify_switch_constructor_kwargs, verify_switch_constructor_args, \
     switch_kwargs, switch_args
-from zepben.evolve import Jumper
+from zepben.evolve.model.cim.iec61970.base.wires.jumper import Jumper
 
 jumper_kwargs = switch_kwargs
 jumper_args = switch_args